fix chef_s3_tests.erl
if i understood the tests correctly, and understood what was being
tested, there was some impedence mismatch between how the old mini_s3
did things and how the new erlcloud and mini_s3 wrapper does things. old
mini_s3 config apparently stores a whole url, either with port
(scheme://whatever.com:port) or without port (scheme://whatever.com).
erlcloud config stores the url in seperate pieces (scheme, uri, port,
etc), and it always stores a port. the issue with these tests was,
the test result wanted the url which may or may not have originally
included a port. this was fine for old mini_s3, as it just stored the
whole url (which either included or didn't include the port), and could
produce this url on demand. but this wasn't fine for erlcloud, because
it always has a port stored, and defaults to a port if there isn't one.
this fix attempted to address this issue.

Signed-off-by: Lincoln Baker <lbaker@chef.io>